Cleanup: Unused code deleted.
