+++ app/src/main/java/ua/maclaren99/pi_android_school_autumn_2020/data/database/AppDatabase.kt
@@ -6,21 +6,16 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 
 
-// Annotates class to be a Room Database with a table (entity) of the Word class
 @Database(entities = arrayOf(Picture::class, User::class, Request::class), version = 1, exportSchema = false)
-public abstract class AppDatabase : RoomDatabase() {
+abstract class AppDatabase : RoomDatabase() {
 
     abstract fun pictureDAO(): PictureDAO
 
     companion object {
-        // Singleton prevents multiple instances of database opening at the
-        // same time.
         @Volatile
         private var INSTANCE: AppDatabase? = null
 
         fun getDatabase(context: Context): AppDatabase {
-            // if the INSTANCE is not null, then return it,
-            // if it is, then create the database
             return INSTANCE ?: synchronized(this) {
                 val instance = Room.databaseBuilder(
                     context.applicationContext,
@@ -28,7 +23,6 @@ public abstract class AppDatabase : RoomDatabase() {
                     "app_database"
                 ).build()
                 INSTANCE = instance
-                // return instance
                 instance
             }
         }
