+++ app/src/main/java/ua/maclaren99/pi_android_school_autumn_2020/ui/MainActivity/PhotoUrlListAdapter.kt
@@ -10,20 +10,18 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import kotlinx.android.synthetic.main.item_photo_card.view.*
 import ua.maclaren99.pi_android_school_autumn_2020.R
-import ua.maclaren99.pi_android_school_autumn_2020.data.network.displayWebViewActivity
+import ua.maclaren99.pi_android_school_autumn_2020.util.displayWebViewActivity
 
 
 class PhotoUrlListAdapter : RecyclerView.Adapter<PhotoUrlListAdapter.PhotoUrlViewHolder>() {
 
     private val photoItemList: MutableList<String> = mutableListOf<String>()
 
-    class PhotoUrlViewHolder(view: View) : RecyclerView.ViewHolder(view)/*, View.OnClickListener*/ {
+    class PhotoUrlViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val urlTextView: TextView = view.url_text_view
         val photoPreview: ImageView = view.photo_preview
         val innerLayout = view.item_photo_card_clayout
         val closeButton = view.close_button
-
-
     }
 
 
@@ -50,8 +48,6 @@ class PhotoUrlListAdapter : RecyclerView.Adapter<PhotoUrlListAdapter.PhotoUrlVie
             .diskCacheStrategy(DiskCacheStrategy.RESOURCE)
             .centerCrop()
             .into(holder.photoPreview)
-
-
     }
 
     override fun getItemCount(): Int = photoItemList.size
@@ -78,14 +74,6 @@ class PhotoUrlListAdapter : RecyclerView.Adapter<PhotoUrlListAdapter.PhotoUrlVie
             photoItemList.removeAt(it)
         }
         notifyDataSetChanged()
-    }
-
-    class PhotoItem(){
-        //TODO("")
-    }
-
-    class RequestHeaderItem(){
-
     }
 
 }
